Add first draft of dynamic auth by the router
Before this modification the router allowed anonymous connection and
provide all the availables method to all clients. This commit is a poc
to authentification managed by the app and allow some rpc call only for
listeners.

This feature is not actually fully implemented as we need to manage more
roles (as actuator or admin/manager) and try to use other authentication
methods.