Remove actuator manager model to just keep actuators
At the begin actuator was considered as channel of actuator manager
which are used as a device managing multiple actuators. To keep things
simple, actuator and actuator manager has been merged in one model.

Actuator will now register procedure to the router as follow :

com.betamachine.actuator.NAME.CHANNEL.LOW|HIGH|TOGGLE

where name refer as the actuator name, channel as a string which for
example refer as the GPIO pin for the ``actuator-rpig`` project and an
action that can be LOW, HIGH or TOGGLE.